requirements.txt
absl-py==2.1.0
addict==2.4.0
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
ConfigArgParse==1.7
constantly==23.10.4
contourpy==1.3.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
exceptiongroup==1.2.2
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.15.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.55.8
fsspec==2024.6.0
gast==0.4.0
google-api-core==2.24.2
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-vision==3.10.1
google-pasta==0.2.0
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.23.4
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.6.1
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.3.1
itsdangerous==2.2.0
jax==0.5.0
jaxlib==0.5.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.2.2
Mako==1.3.9
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.20
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.24.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.3.0
optree==0.11.0
packaging==24.1
pandas==2.2.2
parsel==1.9.1
pillow==11.1.0
pip-chill==1.0.3
platformdirs==4.3.6
plotly==6.0.0
Protego==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
pygame==2.5.2
Pygments==2.18.0
pyodbc==5.2.0
pyOpenSSL==24.1.0
pyparsing==3.2.1
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-tds==1.16.0
pytz==2024.1
PyYAML==6.0.1
queuelib==1.7.0
referencing==0.36.2
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
Scrapy==2.11.2
sentencepiece==0.2.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
SQLAlchemy==2.0.39
sqlalchemy_pytds==1.0.2
starlette==0.46.1
sympy==1.13.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.37.0
tensorflow-macos==2.13.0
termcolor==2.4.0
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.19.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.4
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers@b38612d31263294bc997a21e4aceb1485414fa2d
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.34.0
w3lib==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.21.0
zope.interface==6.4.post2
